requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-docx>=1.1
langchain-core==0.3.*
langchain-huggingface==0.3.*
langchain-chroma>=0.4.2
langchain-text-splitters==0.3.*
langchain-openai>=0.3
transformers==4.52.*
huggingface-hub>=0.33
sentence-transformers>=2.6